refactor: move StatScope to the constants directory
