+++ client/src/modules/CourseStatistics/constants.ts
