--- client/src/modules/CourseStatistics/entities/index.ts
@@ -1,1 +1,0 @@
-export * from './enums';
